requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
requests==2.32.3
aiofiles==24.1.0
python-multipart==0.0.20
python-dotenv==1.0.1
Pillow==10.1.0
APScheduler==3.10.4